api/requirements.txt
fastapi
python-multipart
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
requests